updates for canvases and rendering
